pyproject.toml
[tool.poetry]
name = "streamlit-apps"
version = "0.1.0"
description = ""
authors = ["dcaribou <davidcereijo@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
jupyter = "^1.0.0"
pandas = "^2.1.1"
matplotlib = "^3.8.0"
numpy-financial = "^1.0.0"
streamlit = "^1.27.2"
google-api-python-client = "^2.125.0"
google-auth-oauthlib = "^1.2.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"
watchdog = "^3.0.0"
pytest = "^7.4.3"

# https://docs.pytest.org/en/latest/reference/reference.html#command-line-flags
[tool.pytest.ini_options]
testpaths = [
    "utils",
]
pythonpath = [
    ".",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
